+++ src/peakrdl_rust/templates/src/lib.rs
@@ -3,6 +3,7 @@
 pub mod access;
 #[cfg(not(doctest))]
 pub mod components;
+pub mod encode;
 {% if ctx.has_fixedpoint %}
 pub mod fixedpoint;
 {% endif %}